# Obsevatoire_de_recherche/src/Backend/laravel10/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "This is the API documentation for my application.",
        "contact": {
            "email": "your-email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API server"
        }
    ],
    "paths": {
...
        "/api/ressources/collaborateurs": {
            "get": {
                "tags": [
                    "Collaborators"
                ],
                "summary": "Get list of all collaborators",
                "operationId": "bfa0b07a8cf7f4696a1f05b9fae713b8",
                "responses": {
                    "200": {
                        "description": "A list of collaborators",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TblCollaborateur"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Collaborators"
                ],
                "summary": "Create a new collaborator",
                "operationId": "0623b868a887330fc39b88af8c8120d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TblCollaborateur"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Collaborator created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TblCollaborateur"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/ressources/collaborateurs/{id}": {
            "get": {
                "tags": [
                    "Collaborators"
                ],
                "summary": "Get a collaborator by ID",
                "operationId": "dbd1782b356813fd699301ca2e0d19ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Collaborator details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TblCollaborateur"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Collaborator not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Collaborators"
                ],
                "summary": "Update a collaborator",
                "operationId": "38801022244d0a79cf11a1ff41855127",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TblCollaborateur"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Collaborator updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TblCollaborateur"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Collaborator not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Collaborators"
                ],
                "summary": "Delete a collaborator",
                "operationId": "8f36d658e94e7faabd9c66be07b8da52",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Collaborator deleted"
                    },
                    "404": {
                        "description": "Collaborator not found"
                    }
                }
            }
        },
        "/api/ressources/documents": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get list of all documents",
                "operationId": "e059be3fabb6af1ccf88067a25b3a480",
                "responses": {
                    "200": {
                        "description": "A list of documents",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TblDocument"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Create a new document",
                "operationId": "698daea28a49ea4b8b0fea05979507ae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TblDocument"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Document created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TblDocument"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/ressources/documents/{id}": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get a document by ID",
                "operationId": "0f4fc7603b3df3b74b4dd6a179c227a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TblDocument"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Document not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Documents"
                ],
                "summary": "Update a document",
                "operationId": "b3457c185bb0af57181739041cc79dfa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TblDocument"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Document updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TblDocument"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Document not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Documents"
                ],
                "summary": "Delete a document",
                "operationId": "b09ff65c58c39ad0503652192a5669d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Document deleted"
                    },
                    "404": {
                        "description": "Document not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TblCollaborateur": {
                "title": "TblCollaborateur",
                "required": [
                    "nom_collab",
                    "email_collab"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the collaborator",
                        "type": "integer",
                        "format": "int64"
                    },
                    "nom_collab": {
                        "description": "Name of the collaborator",
                        "type": "string"
                    },
                    "email_collab": {
                        "description": "Email of the collaborator",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "TblDocument": {
                "title": "TblDocument",
                "required": [
                    "nom_doc",
                    "lien_doc",
                    "type_doc",
                    "resume",
                    "tbl_projet_id"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the document",
                        "type": "integer",
                        "format": "int64"
                    },
                    "nom_doc": {
                        "description": "Name of the document",
                        "type": "string"
                    },
                    "lien_doc": {
                        "description": "Link to the document",
                        "type": "string"
                    },
                    "type_doc": {
                        "description": "Type of the document (PDF, WORD, POWEPOINT)",
                        "type": "string"
                    },
                    "resume": {
                        "description": "Summary of the document",
                        "type": "string"
                    },
                    "tbl_projet_id": {
                        "description": "ID of the related project",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}